Refactor i18n calendar formatter
