+++ react/features/base/i18n/dateUtil.js
@@ -2,13 +2,11 @@
 
 import moment from 'moment';
 
-import { i18next } from '../i18n';
+import i18next from './i18next';
 
 // MomentJS uses static language bundle loading, so in order to support dynamic
 // language selection in the app we need to load all bundles that we support in
 // the app.
-// FIXME: If we decide to support MomentJS in other features as well we may need
-// to move this import and the lenient matcher to the i18n feature.
 require('moment/locale/bg');
 require('moment/locale/de');
 require('moment/locale/eo');
@@ -32,7 +30,7 @@ require('moment/locale/zh-cn');
 
 /**
  * Returns a localized date formatter initialized with a specific {@code Date}
- * or time stamp ({@code number}).
+ * or timestamp ({@code number}).
  *
  * @private
  * @param {Date | number} dateOrTimeStamp - The date or unix timestamp (ms)
