+++ react/features/calendar-sync/components/ConferenceNotification.native.js
@@ -9,10 +9,9 @@ import { connect } from 'react-redux';
 
 import { appNavigate } from '../../app';
 import { getURLWithoutParams } from '../../base/connection';
-import { translate } from '../../base/i18n';
+import { getLocalizedDateFormatter, translate } from '../../base/i18n';
 import { Icon } from '../../base/font-icons';
 import { ASPECT_RATIO_NARROW } from '../../base/responsive-ui';
-import { getLocalizedDateFormatter } from '../../base/util';
 
 import styles from './styles';
